operations: Replace use of Vector with a unsynchronized equivalent.
Early classes of the Java API, such as Vector, Hashtable and StringBuffer, were synchronized to make them thread-safe. Unfortunately, synchronization has a big negative impact on performance, even when using these collections from a single thread.

It is better to use their new unsynchronized replacements. ArrayList is pretty much a drop-in replacement for Vectors and I cant see any reason to use Vector unless there is some sort of special requirement for it. ArrayList will generally give you better performance and only half the data growth when items are added to the list.